refactor(icons): move devicons stuff to util.mini
